requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==3.2.3
djangorestframework==3.12.4
drf-spectacular==0.27.2
exceptiongroup==1.2.2
flake8==6.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.3.0
